requirements.txt
uvicorn==0.15.0
fastapi==0.70.0
pydantic[email]==1.10.7
google-cloud-storage==2.9.0
python-dotenv==1.0.0
PyJWT==1.7.1
python-decouple==3.3
mysql-connector==2.2.9
python-multipart==0.0.6
Pillow==9.5.0
numpy==1.24.3
cloud-sql-python-connector
streamlit
cryptography
fastapi_pagination
pandas
scikit-learn
tensorflow